+++ R/io.R
@@ -61,7 +61,14 @@ file_string = function(file) {
   raw_string(x)
 }
 
-# read all records of a binary file as a raw vector by default
+#' Read all records of a binary file as a raw vector by default
+#'
+#' This is a wrapper function of \code{\link{readBin}()} with default arguments
+#' \code{what = "raw"} and \code{n = \link{file.size}(file)}, which means it
+#' will read the full content of a binary file as a raw vector by default.
+#' @param file,what,n,... Arguments to be passed to \code{readBin()}.
+#' @return A vector returned from \code{readBin()}.
+#' @export
 read_bin = function(file, what = 'raw', n = file.info(file)$size, ...) {
   readBin(file, what, n, ...)
 }
@@ -221,4 +228,34 @@ github_releases = function(repo, subpath = '', pattern = '(v[0-9.]+)') {
   r = sprintf('^.*?releases/tag/%s".*', pattern)
   v = gsub(r, '\\1', grep(r, h, value = TRUE))
   unique(v)
+}
+
+#' Generate a message with \code{cat()}
+#'
+#' This function is similar to \code{\link{message}()}, and the difference is
+#' that \code{msg_cat()} uses \code{\link{cat}()} to write out the message,
+#' which is sent to \code{\link{stdout}} instead of \code{\link{stderr}}. The
+#' message can be suppressed by \code{\link{suppressMessages}()}.
+#' @param ... Character strings of messages, which will be concatenated into one
+#'   string via \code{paste(c(...), collapse = '')}.
+#' @note By default, a newline will not be appended to the message. If you need
+#'   a newline, you have to explicitly add it to the message (see
+#'   \sQuote{Examples}).
+#' @return Invisible \code{NULL}, with the side-effect of printing the message.
+#' @seealso This function was inspired by \code{rlang::inform()}.
+#' @export
+#' @examples
+#' {
+#' # a message without a newline at the end
+#' xfun::msg_cat('Hello world!')
+#' # add a newline at the end
+#' xfun::msg_cat(' This message appears right after the previous one.\n')
+#' }
+#' suppressMessages(xfun::msg_cat('Hello world!'))
+msg_cat = function(...) {
+  x = paste(c(...), collapse = '')
+  withRestarts({
+    signalCondition(simpleMessage(x))
+    cat(x)
+  }, muffleMessage = function() invisible(NULL))
 }